jsonschema2sql: put `SqlQuery` class in its own file
